More work on ColorPickerButton widget and mixin
A tooltip is always generated for the button, with the click instructions. The instructions are dynamic depending on the option to use default color picker.
Updated flow for the personality traits + we now color the text with the custom color.
Added fading animation of the button's highlight color.
Removed the blink animation.
Fixed issue from using :GetRGB() instead of :GetRGBAsBytes() when displaying the color pickers.